Commit before Pool Branch.
I am not fully comfortable with branches. In theory they are easy to understand but in practice I'm still a little in the dark. Also, based on past experiences, the pool manager is fairly involved with all the things. So seems like a good time to practice.

This commit includes some testing things for spawning.