Return
Oh. It was easier than I thought to do the return object. Seems to be working as intended so far. I guess if we can pool things now, I could keep it simple and start working on waves / subscriptions? Although I kind of want to do more features like dynamically create pools / empty objects in the hierarchy for unity. But probably outside the scope of the project I'm trying to do.