Very Cool Movement
Added some very very cool movement. The enemies now move along the Z axis until it reaches the wall, Then the coroutine dies and we don't have to do a bunch of 'If" checks in Update(). Seems smarter to use coroutines. Helps keep it super lean since their could be a gazillion enemies on screen at once. Imagien if all of them had to do like 10 if checks everytime. Way faster to just move precisely when needed and no more.