Attacking
Enemy "attack" in. Added event system so everyone knows to stop attacking. Did the 'correct' thing and made sure to unsubscribe to the events if disabled. Stopped all coroutines so I think that covers everything. Maybe not. Hopefully. Anyway, mostly hooked up to UI. Need to get the slider in there and I'll do that when I return.