Restart Flow
Got The Director to be initialized first. Weird. ANyway, I have the barebones of the restart. I have each script handling it's very own restarting things. That will probably require more coding things to be able to do that but I think it'll be good. At the minimum everything is subscribed to it and we fire off a debug.log when it happens. Gooooo Team!