Omg this Code Review is not really a code review
This is now a bug fixing endeavor where I fix what I broke while moving some things and doing a code review. All easy fixes. But now I'm satisfied with the animator and it's behavior. Next is actually scripting the triggers and communicating things to TheDirector.  Once triggers done, I think that'll be all the logic I need for the UI manager on restarting. Then I can move to Wave Manager. That'll be cool. And I must not forget the player scripts and things!