More Animator Work
Hooked up a couple of buttons to test the fadetoblack and fadeblacktoclear. Also renamed FadeIn and FadeOut as it is ambiguous. I need to fix the time issue with the pause menu as I set timescale to 0 when paused. Maybe I really do need to switch to DOTween instead of animators.