Alright. Animator -> DOTween
I'm switching to DOTween I guess. Got rid of animator and now It fades to black. It isn't fully complete because we need to fade to black, wait for things then fade from black to clear. That's another day though. I think the UI logic is complete though. Time for Code Review!